Implement re-connect and re-subscribe on the OPC UA client adapter
This change does implement a functionality that the OPC UA client
periodically tries to re-connect to the OPC UA server in case
of the initial connection. In that first startup case the Milo client
does not yet re-establish the connection.

Also will the adapter now try to re-register items if they failed
to register in the first time.

The test applications are now also available in a "split up" version
in order to test re-starting client and server.

The server side fixes an implementation in the namespace which actually
prevented write requests.